requirements.txt
numpy==1.23.3
openai==1.79.0
Pillow==9.0.1
Pillow==11.2.1
scikit_learn==1.6.1
torch==2.5.1+cu118
tqdm==4.57.0
transformers==4.51.3
